+++ tools/send-to-caas/bulkpublish.css
@@ -1,0 +1,24 @@
+#bulkpublish {
+  margin: 20px;
+}
+
+.bulk-publisher {
+  width: 500px;
+}
+
+.bulk-publisher .title {
+  grid-template-columns: 1fr;
+}
+
+dd.content .field {
+  padding: 0 20px;
+  margin: 20px 0
+}
+
+.results {
+  display: flex;
+}
+
+.results > div {
+  padding: 20px;
+}
