+++ tools/send-to-caas/send-to-caas.js
@@ -1,20 +1,19 @@
 /* eslint-disable new-cap */
 /* global tingle */
 /* eslint-disable no-alert */
-import getUuid from '../../libs/utils/getUuid.js';
+
 import { loadScript, loadStyle } from '../../libs/utils/utils.js';
+import {
+  getCardMetadata,
+  getCaasProps,
+  getImsToken,
+  isPagePublished,
+  loadCaasTags,
+  postDataToCaaS,
+  setConfig,
+} from './send-utils.js';
 
 loadStyle('https://milo.adobe.com/tools/send-to-caas/send-to-caas.css');
-
-const CAAS_TAG_URL = 'https://www.adobe.com/chimera-api/tags';
-const HLX_ADMIN_STATUS = 'https://admin.hlx.page/status';
-const IMS_CLIENT_ID = 'milo_ims';
-const IMS_PROD_URL = 'https://auth.services.adobe.com/imslib/imslib.min.js';
-const URL_POSTXDM = 'https://14257-milocaasproxy-stage.adobeio-static.net/api/v1/web/milocaas/postXDM';
-const VALID_URL_RE = /(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_+.~#?&//=]*)/;
-
-const isKeyValPair = /(\s*\S+\s*:\s*\S+\s*)/;
-const isValidUrl = (u) => VALID_URL_RE.test(u);
 
 const [setPublishingTrue, setPublishingFalse, isPublishing] = (() => {
   let publishing = false;
@@ -24,281 +23,6 @@ const [setPublishingTrue, setPublishingFalse, isPublishing] = (() => {
     () => publishing,
   ];
 })();
-
-const [setConfig, getConfig] = (() => {
-  let config = {};
-  return [
-    (c) => {
-      config = { ...config, ...c };
-      return config;
-    },
-    () => config,
-  ];
-})();
-
-const getKeyValPairs = (s) => {
-  if (!s) return [];
-  return s
-    .split(',')
-    .filter((v) => v.length)
-    .filter((v) => isKeyValPair.test(v))
-    .map((v) => {
-      const [key, ...value] = v.split(':');
-      return {
-        key: key.trim(),
-        value: value.join(':').trim(),
-      };
-    });
-};
-
-const addHost = (url) => {
-  if (url.startsWith('http')) return url;
-  const { host } = getConfig();
-  return `https://${host}${url.startsWith('/') ? '' : '/'}${url}`
-};
-
-const getMetaContent = (propType, propName) => {
-  const metaEl = document.querySelector(`meta[${propType}='${propName}']`);
-  if (!metaEl) return undefined;
-  return metaEl.content;
-};
-
-const prefixHttps = (url) => {
-  if (!(url?.startsWith('https://') || url?.startsWith('http://'))) {
-    return `https://${url}`;
-  }
-  return url;
-};
-
-const checkUrl = (url, errorMsg) => {
-  if (url === undefined) return url;
-  return isValidUrl(url) ? prefixHttps(url) : { error: errorMsg };
-};
-
-// Case-insensitive search through tag name, path, id and title for the searchStr
-const findTag = (tags, searchStr, ignore = []) => {
-  const childTags = [];
-  let matchingTag = Object.values(tags).find((tag) => {
-    if (
-      ignore.includes(tag.title)
-      || ignore.includes(tag.name)
-      || ignore.includes(tag.path)
-      || ignore.includes(tag.tagID)
-    ) return false;
-
-    if (tag.tags && Object.keys(tag.tags).length) {
-      childTags.push(tag.tags);
-    }
-
-    const tagMatches = [
-      tag.title.toLowerCase(),
-      tag.name,
-      tag.path,
-      tag.path.replace('/content/cq:tags/', ''),
-      tag.tagID,
-    ];
-
-    if (tagMatches.includes(searchStr.toLowerCase())) return true;
-
-    return false;
-  });
-
-  if (!matchingTag) {
-    childTags.some((childTag) => {
-      matchingTag = findTag(childTag, searchStr, ignore);
-      return matchingTag;
-    });
-  }
-
-  return matchingTag;
-};
-
-const fetchCaasTags = async () => {
-  try {
-    const resp = await fetch(CAAS_TAG_URL);
-    if (resp.ok) {
-      const json = await resp.json();
-      return json.namespaces.caas.tags;
-    }
-  } catch (e) {
-    // ignore
-  }
-
-  const { default: caasTags } = await import('../../libs/blocks/caas-config/caas-tags.js');
-  return caasTags.namespaces.caas.tags;
-};
-
-const [getCaasTags, loadCaasTags] = (() => {
-  let tags;
-  return [
-    () => tags,
-    async () => {
-      try {
-        const resp = await fetch(CAAS_TAG_URL);
-        if (resp.ok) {
-          const json = await resp.json();
-          tags = json.namespaces.caas.tags;
-        }
-      } catch (e) {
-        // ignore
-      }
-
-      const { default: caasTags } = await import('../../libs/blocks/caas-config/caas-tags.js');
-      tags = caasTags.namespaces.caas.tags;
-    }];
-})();
-
-const getTag = (tagName, errors) => {
-  if (!tagName) return undefined;
-  const caasTags = getCaasTags();
-  // search all except Events first
-  const tag = findTag(caasTags, tagName, ['Events']) || findTag(caasTags.events.tags, tagName, []);
-
-  if (!tag) {
-    errors.push(tagName);
-  }
-
-  return tag;
-};
-
-const getTags = (s) => {
-  let rawTags = [];
-  if (s) {
-    rawTags = s.split(',').map((t) => t.trim());
-  } else {
-    rawTags = [...document.querySelectorAll("meta[property='article:tag']")].map(
-      (metaEl) => metaEl.content,
-    );
-  }
-
-  const errors = [];
-
-  if (!rawTags.length) rawTags = ['Article']; // default if no tags found
-
-  const tagIds = rawTags.map((tag) => getTag(tag, errors))
-    .filter((tag) => tag !== undefined)
-    .map((tag) => tag.tagID);
-
-  const tags = [...new Set(tagIds)]
-    .map((tagID) => ({ id: tagID }));
-
-  return {
-    tagErrors: errors,
-    tags,
-  };
-};
-
-const getDateProp = (dateStr, errorMsg) => {
-  if (!dateStr) return undefined;
-  try {
-    const date = new Date(dateStr);
-    if (date.getFullYear() < 2000) return { error: `${errorMsg} - Date is before the year 2000` };
-    return date.toISOString();
-  } catch (e) {
-    return { error: errorMsg };
-  }
-};
-
-const getOrigin = () => {
-  const origin = getConfig().project || getConfig().repo;
-  if (origin) return origin;
-
-  if (window.location.hostname.endsWith('.hlx.page')) {
-    const [, repo] = window.location.hostname.split('.')[0].split('--');
-    return repo;
-  }
-
-  throw new Error('No Project or Repo defined in config');
-};
-
-const getImagePathMd = (keyName) => {
-  const mdEl = document.querySelector('.card-metadata');
-  let url = '';
-  [...mdEl.children].some(n => {
-    const key = n.firstElementChild.textContent?.trim().toLowerCase();
-    if (key !== keyName) return false;
-
-    const img = n.lastElementChild.querySelector('img');
-
-    if (img) {
-      url = new URL(img.src)?.pathname;
-    } else {
-      url = n.lastElementChild.textContent?.trim();
-    }
-    return true;
-  });
-  return url;
-}
-
-const getThumbnailUrl = () => {
-  const thumbUrl = getImagePathMd('cardimagepath')
-    || document.querySelector('meta[property="og:image"]')?.content
-    || document.querySelector('main')?.querySelector('img')?.src;
-
-  if (!thumbUrl) return null;
-  return addHost(thumbUrl);
-};
-
-const getFirstImageAlt = () => document.querySelector('main')?.querySelector('img')?.alt;
-
-const getBadges = (s) => {
-  let hasImageBadge = false;
-  const keyValPairs = getKeyValPairs(s)
-    .map((pair) => {
-      const type = pair.key?.toLowerCase();
-      let value = pair.value;
-      if (type === 'image' && value) {
-        hasImageBadge = true;
-        value = addHost(value);
-      }
-      return (type && value)
-      ? ({ type, value })
-      : null;
-    })
-    .filter(i => i !== null);
-
-  if (!hasImageBadge) {
-    const imgPath = getImagePathMd('badgeimage');
-    if (imgPath) {
-      keyValPairs.push({ type: 'image', value: addHost(imgPath) });
-    }
-  }
-  return keyValPairs;
-}
-
-const getImsToken = async () => {
-  window.adobeid = {
-    client_id: IMS_CLIENT_ID,
-    environment: 'prod',
-    scope: 'AdobeID,openid',
-  };
-
-  if (!window.adobeIMS) {
-    await loadScript(IMS_PROD_URL);
-  }
-  return window.adobeIMS?.getAccessToken()?.token;
-};
-
-const isPagePublished = async () => {
-  let { branch, repo, owner } = getConfig();
-  if (!(branch || repo || owner)
-    && window.location.hostname.endsWith('.hlx.page')) {
-    [branch, repo, owner] = window.location.hostname.split('.')[0].split('--');
-  }
-
-  if (!(branch || repo || owner)) {
-    throw new Error(`Branch, Repo or Owner is not set - branch: ${branch}, repo: ${repo}, owner: ${owner}`);
-  }
-
-  const res = await fetch(
-    `${HLX_ADMIN_STATUS}/${owner}/${repo}/${branch}${window.location.pathname}`,
-  );
-  if (res.ok) {
-    const json = await res.json();
-    return json.live.status === 200;
-  }
-  return false;
-};
 
 // Tingle is the js library for displaying modals
 const loadTingleModalFiles = async () => {
@@ -310,7 +34,7 @@ const loadTingleModalFiles = async () => {
   }
 };
 
-const showAlert = (msg, { error = false, onClose } = {}) => {
+const showAlert = (msg, { error = false, onClose, showBtn = true, btnText = 'OK' } = {}) => {
   const modal = new tingle.modal({
     footer: true,
     closeMethods: ['overlay', 'escape'],
@@ -324,8 +48,11 @@ const showAlert = (msg, { error = false, onClose } = {}) => {
     msgContent = `<div class="modal-error"><div class="modal-alert"></div><div>${msg}</div></div>`;
   }
   modal.setContent(msgContent);
-  modal.addFooterBtn('OK', 'tingle-btn tingle-btn--primary tingle-btn--pull-right', () => modal.close());
+  if (showBtn) {
+    modal.addFooterBtn(btnText, 'tingle-btn tingle-btn--primary tingle-btn--pull-right', () => modal.close());
+  }
   modal.open();
+  return modal;
 };
 
 const showConfirm = (msg, {
@@ -368,192 +95,6 @@ const showConfirm = (msg, {
   }
   modal.open();
 });
-
-/** card metadata props - either a func that computes the value or
- * 0 to use the string as is
- * funcs that return an object with { error: string } will report the error
- */
-const props = {
-  arbitrary: (s) => getKeyValPairs(s).map((pair) => ({ key: pair.key, value: pair.value })),
-  badges: getBadges,
-  bookmarkaction: 0,
-  bookmarkenabled: (s = '') => {
-    if (s) {
-      const lcs = s.toLowerCase();
-      if (lcs === 'true' || lcs === 'on' || lcs === 'yes') {
-        return true;
-      }
-    }
-    return undefined;
-  },
-  bookmarkicon: 0,
-  contentid: (_, options) => getUuid(options.prodUrl),
-  contenttype: (s) => s || getMetaContent('property', 'og:type') || 'Article',
-  // TODO - automatically get country
-  country: (s) => s || 'us',
-  created: (s) => {
-    if (s) {
-      return getDateProp(s, `Invalid Created Date: ${s}`);
-    }
-
-    const pubDate = getMetaContent('name', 'publication-date');
-    return pubDate
-      ? getDateProp(pubDate, `publication-date metadata is not a valid date: ${pubDate}`)
-      : getDateProp(document.lastModified, `document.lastModified is not a valid date: ${document.lastModified}`);
-  },
-  cta1icon: (s) => checkUrl(s, `Invalid Cta1Icon url: ${s}`),
-  cta1style: 0,
-  cta1text: 0,
-  cta1url: (s, options) => {
-    const url = s || options.prodUrl || window.location.origin + window.location.pathname;
-    return checkUrl(url, `Invalid Cta1Url: ${url}`);
-  },
-  cta2icon: (s) => checkUrl(s, `Invalid Cta2Icon url: ${s}`),
-  cta2style: 0,
-  cta2text: 0,
-  cta2url: (s) => checkUrl(s, `Invalid Cta2Url: ${s}`),
-  description: (s) => s || getMetaContent('name', 'description') || '',
-  details: 0,
-  entityid: (_, options) => getUuid(options.prodUrl),
-  env: (s) => s || '',
-  eventduration: 0,
-  eventend: (s) => getDateProp(s, `Invalid Event End Date: ${s}`),
-  eventstart: (s) => getDateProp(s, `Invalid Event Start Date: ${s}`),
-  floodgatecolor: (s) => s || 'default',
-  headline: 0,
-  // TODO: automatically get lang
-  lang: (s) => s || 'en',
-  modified: (s) => (s
-    ? getDateProp(s, `Invalid Modified Date: ${s}`)
-    : getDateProp(document.lastModified, `document.lastModified is not a valid date: ${document.lastModified}`)),
-  origin: (s) => s || getOrigin(),
-  playurl: (s) => checkUrl(s, `Invalid PlayURL: ${s}`),
-  primarytag: (s) => {
-    const tag = getTag(s);
-    return tag ? { id: tag.tagID } : {};
-  },
-  style: (s) => s || 'default',
-  tags: (s) => getTags(s),
-  thumbalt: (s) => s || getFirstImageAlt(),
-  thumburl: (s) => (s ? checkUrl(s, `Invalid Thumbnail URL: ${s}`) : getThumbnailUrl()),
-  title: (s) => s || getMetaContent('property', 'og:title') || '',
-  uci: (s) => s || window.location.pathname,
-  url: (s, options) => {
-    const url = s || options.prodUrl || window.location.origin + window.location.pathname;
-    return checkUrl(url, `Invalid URL: ${url}`);
-  },
-};
-
-// Map the flat props into the structure needed by CaaS
-const getCaasProps = (p) => {
-  const caasProps = {
-    entityId: p.entityid,
-    contentId: p.contentid,
-    contentType: p.contenttype,
-    environment: p.env,
-    url: p.url,
-    floodGateColor: p.floodgatecolor,
-    universalContentIdentifier: p.uci,
-    title: p.title,
-    description: p.description,
-    createdDate: p.created,
-    modifiedDate: p.modified,
-    tags: p.tags,
-    primaryTag: p.primarytag,
-    ...(p.thumburl && {
-      thumbnail: {
-        altText: p.thumbalt,
-        url: p.thumburl,
-      },
-    }),
-    country: p.country,
-    language: p.lang,
-    cardData: {
-      style: p.style,
-      headline: p.headline || p.title,
-      ...(p.details && { details: p.details }),
-      ...((p.bookmarkenabled || p.bookmarkicon || p.bookmarkaction) && {
-        bookmark: {
-          enabled: p.bookmarkenabled,
-          bookmarkIcon: p.bookmarkicon,
-          action: p.bookmarkaction,
-        },
-      }),
-      badges: p.badges,
-      ...(p.playurl && { playUrl: p.playurl }),
-      cta: {
-        primaryCta: {
-          text: p.cta1text,
-          url: p.cta1url,
-          style: p.cta1style,
-          icon: p.cta1icon,
-        },
-        ...(p.cta2url && {
-          secondaryCta: {
-            text: p.cta2text,
-            url: p.cta2url,
-            style: p.cta2style,
-            icon: p.cta2icon,
-          },
-        }),
-      },
-      ...((p.eventduration || p.eventstart || p.eventend) && {
-        event: {
-          duration: p.eventduration,
-          startDate: p.eventstart,
-          endDate: p.eventend,
-        },
-      }),
-    },
-    origin: p.origin,
-    ...(p.arbitrary?.length && { arbitrary: p.arbitrary }),
-  };
-  return caasProps;
-};
-
-const getCaaSMetadata = async (pageMd, options) => {
-  const md = {};
-  const errors = [];
-  let tagErrors = [];
-  let tags = [];
-  // for-of required to await any async computeVal's
-  // eslint-disable-next-line no-restricted-syntax
-  for (const [key, computeFn] of Object.entries(props)) {
-    // eslint-disable-next-line no-await-in-loop
-    const val = computeFn ? await computeFn(pageMd[key], options) : pageMd[key];
-    if (val?.error) {
-      errors.push(val.error);
-    } else if (val?.tagErrors !== undefined) {
-      tagErrors = val.tagErrors;
-      md[key] = val.tags;
-      tags = val.tags.map((t) => t.id);
-    } else if (val !== undefined) {
-      md[key] = val;
-    }
-  }
-
-  return { caasMetadata: md, errors, tags, tagErrors };
-};
-
-const parseCardMetadata = () => {
-  const pageMd = {};
-  const mdEl = document.querySelector('.card-metadata');
-  if (mdEl) {
-    mdEl.childNodes.forEach((n) => {
-      const key = n.children?.[0]?.textContent?.toLowerCase();
-      const val = n.children?.[1]?.textContent?.toLowerCase();
-      if (!key) return;
-
-      pageMd[key] = val;
-    });
-  }
-  return pageMd;
-};
-
-const getCardMetadata = async (options) => {
-  const pageMd = parseCardMetadata();
-  return getCaaSMetadata(pageMd, options);
-};
 
 const displayPublishingModal = () => {
   const publishingModal = new tingle.modal({
@@ -708,50 +249,30 @@ const checkIms = async (publishingModal) => {
 };
 
 const postToCaaS = async ({ accessToken, caasEnv, caasProps, draftOnly, publishingModal }) => {
-  const options = {
-    method: 'POST',
-    body: JSON.stringify(caasProps),
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      draft: draftOnly,
-      'caas-env': caasEnv,
-    },
-  };
-
   try {
-    const res = await fetch(URL_POSTXDM, options);
-    if (res !== undefined) {
-      const text = await res.text();
+    const response = await postDataToCaaS({ accessToken, caasEnv, caasProps, draftOnly });
 
-      let response;
-      try {
-        response = JSON.parse(text);
-      } catch {
-        response = text;
-      }
+    publishingModal.close();
 
-      publishingModal.close();
-
-      if (response.success) {
-        showAlert(
-          `<p>Successfully published page to CaaS!<p><p>Card ID: ${caasProps.entityId}</p>`,
-          { onClose: setPublishingFalse },
-        );
-      } else if (response.error?.startsWith('Invalid User: Not an Adobe employee')) {
-        const msg = 'Please login with your Adobe company account.  Do you want to try logging in again?';
-        const shouldLogIn = await showConfirm(msg, {
-          cancelBtnType: 'grey',
-          cancelText: 'Cancel',
-          ctaText: 'Login',
-        });
-        setPublishingFalse();
-        if (shouldLogIn) window.adobeIMS.signIn();
-      } else {
-        showAlert(
-          response.message || response.error || JSON.stringify(response),
-          { error: true, onClose: setPublishingFalse },
-        );
-      }
+    if (response.success) {
+      showAlert(
+        `<p>Successfully published page to CaaS!<p><p>Card ID: ${caasProps.entityId}</p>`,
+        { onClose: setPublishingFalse },
+      );
+    } else if (response.error?.startsWith('Invalid User: Not an Adobe employee')) {
+      const msg = 'Please login with your Adobe company account.  Do you want to try logging in again?';
+      const shouldLogIn = await showConfirm(msg, {
+        cancelBtnType: 'grey',
+        cancelText: 'Cancel',
+        ctaText: 'Login',
+      });
+      setPublishingFalse();
+      if (shouldLogIn) window.adobeIMS.signIn();
+    } else {
+      showAlert(
+        response.message || response.error || JSON.stringify(response),
+        { error: true, onClose: setPublishingFalse },
+      );
     }
   } catch (e) {
     publishingModal.close();
@@ -763,7 +284,9 @@ const postToCaaS = async ({ accessToken, caasEnv, caasProps, draftOnly, publishi
 const sendToCaaS = async ({ host = '', project = '', branch = '', repo = '', owner = '' } = {}) => {
   if (isPublishing()) return;
 
-  setConfig({ host, project, branch, repo, owner });
+  setConfig({
+    host, project, branch, repo, owner, doc: document,
+  });
 
   setPublishingTrue();
 
@@ -797,3 +320,9 @@ document.addEventListener('send-to-caas', async (e) => {
   const { host, project, branch, repo, owner } = e.detail;
   sendToCaaS({ host, project, branch, repo, owner });
 });
+
+export {
+  loadTingleModalFiles,
+  showAlert,
+  showConfirm,
+};
